+++ packages/quickjs_extensions/lib/src/quickjs_extension_default_store.dart
@@ -1,0 +1,2 @@
+export 'quickjs_extension_default_store_preferences.dart'
+    if (dart.library.io) 'quickjs_extension_default_store_io.dart';